Heat conduction
Nikhil Nayak wrote a solution for this problem.  In my solution the program uses pointers to get the information from the user.  Then, it uses a 'switch' statement to determine which calculation it needs to to in order to solve the head conductdion problem.

Nikhil, or 'nikhilnayak98', posted solutions to the problems from "Problem Solving and Program Design in C" by Jerri R. Hanly and Elliott B. Koffman on GitHub in his 'CSE2041' repository.