Grade and record quiz
This is a program that grades and records quizzes.  It takes information from a file, adds it to arrays, and takes new information from the user--if any, stores it in the arrays and the input file, then prints it to an output file.  Working on this problem gave me practice with coordinating arrays of type int and double with a grid array of characters.

While working on this problem, even getting started on it, I learned from 'nikhilnayak98' on GitHub in his 'CSE2041' repository for the problems in this book that I can deal with the input file one section at a time, instead of writing a loop that would read all of it at once and put all the information in its respective variables or arrays.  So, I owe a special thanks to him.