Add files via upload
Nikhil, or nikhilnayak98, wrote a solution for this problem and most of the problems from "Problem Solving and Program Design in C" by Jerri R. Hanly and Elliott B. Koffman on GitHub in his 'CSE2041' repository.  My solution adds to his by using pointers to get the information from the user.  Then, it uses a 'switch' statement to determine which calculation it needs to to in order to solve the head conduction problem.