Grade and record quizzes
While working on this problem, even getting started on it, I learned from 'nikhilnayak98' on GitHub in his 'CSE2041' repository for the problems in this book that I can deal with the input file one section at a time, instead of writing a loop that would read all of it at once and put all the information in its respective variables or arrays.  So, I owe a special thanks to him.