Create ch 7 program 9, hours
This is a program for a game of literary "horse" that I wrote in response to a textbook problem.  It includes notes about what I learned by writing the program.  I also had an online tutor look at the problem and make suggestions, which made the program better but added a lot of work!