Create ch 7 program 12, trace notes
Here are notes I used to trace the program execution.  They might be helpful for someone interested in how the program works.  It also shows the 'product_ids' array after the program has put the numbers from the 'Product_array.txt' file in order.  With the 'Product_array.txt' file, it is really interesting how the program picks up element 978 from increment 4 and switches it forward until it finds element 998, drops 978, picks up 998 and switches it forward to increment 19, the last increment in the array.  Then, on the next round, after switching other numbers forward, it finds 978 again and moves it into the 18th increment, the second highest position.